# [android] don't probe directories for .config files