--- a/mono/metadata/mono-config.c
+++ b/mono/metadata/mono-config.c
@@ -618,7 +618,10 @@ mono_config_for_assembly_internal (MonoImage *assembly)
 	MONO_REQ_GC_UNSAFE_MODE;
 
 	MonoConfigParseState state = {NULL};
-	int got_it = 0, i;
+#ifndef TARGET_ANDROID
+	int got_it = 0;
+#endif
+	int i;
 	char *aname, *cfg, *cfg_name;
 	const char *bundled_config;
 	
@@ -634,6 +637,7 @@ mono_config_for_assembly_internal (MonoImage *assembly)
 	mono_config_parse_file_with_context (&state, cfg_name);
 	g_free (cfg_name);
 
+#ifndef TARGET_ANDROID
 	cfg_name = g_strdup_printf ("%s.config", mono_image_get_name (assembly));
 	const char *cfg_dir = mono_get_config_dir ();
 	if (!cfg_dir) {
@@ -657,6 +661,7 @@ mono_config_for_assembly_internal (MonoImage *assembly)
 			break;
 	}
 	g_free (cfg_name);
+#endif
 }
 
 /**